+++ src/Resources/MoonShineUserResource.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace MoonShine\Resources;
 
-use Illuminate\Support\Facades\Route;
 use Illuminate\Validation\Rule;
 use MoonShine\Actions\ExportAction;
 use MoonShine\Decorations\Block;
@@ -22,12 +21,14 @@ use MoonShine\Filters\DateFilter;
 use MoonShine\Filters\TextFilter;
 use MoonShine\FormComponents\ChangeLogFormComponent;
 use MoonShine\FormComponents\PermissionFormComponent;
-use MoonShine\Http\Controllers\PermissionController;
 use MoonShine\Models\MoonshineUser;
 use MoonShine\Models\MoonshineUserRole;
+use MoonShine\Traits\Resource\WithUserPermissions;
 
 class MoonShineUserResource extends Resource
 {
+    use WithUserPermissions;
+
     public static string $model = MoonshineUser::class;
 
     public string $titleField = 'name';
@@ -43,61 +44,61 @@ class MoonShineUserResource extends Resource
     {
         return [
             Grid::make([
-               Column::make([
-                   Block::make(trans('moonshine::ui.resource.main_information'), [
-                       ID::make()
-                           ->sortable()
-                           ->useOnImport()
-                           ->showOnExport(),
+                Column::make([
+                    Block::make(trans('moonshine::ui.resource.main_information'), [
+                        ID::make()
+                            ->sortable()
+                            ->useOnImport()
+                            ->showOnExport(),
 
-                       BelongsTo::make(
-                           trans('moonshine::ui.resource.role'),
-                           'moonshine_user_role_id',
-                           new MoonShineUserRoleResource()
-                       )
-                           ->showOnExport(),
+                        BelongsTo::make(
+                            trans('moonshine::ui.resource.role'),
+                            'moonshine_user_role_id',
+                            new MoonShineUserRoleResource()
+                        )
+                            ->showOnExport(),
 
-                       Text::make(trans('moonshine::ui.resource.name'), 'name')
-                           ->required()
-                           ->useOnImport()
-                           ->showOnExport(),
+                        Text::make(trans('moonshine::ui.resource.name'), 'name')
+                            ->required()
+                            ->useOnImport()
+                            ->showOnExport(),
 
-                       Image::make(trans('moonshine::ui.resource.avatar'), 'avatar')
-                           ->removable()
-                           ->showOnExport()
-                           ->disk(config('filesystems.default'))
-                           ->dir('moonshine_users')
-                           ->allowedExtensions(['jpg', 'png', 'jpeg', 'gif']),
+                        Image::make(trans('moonshine::ui.resource.avatar'), 'avatar')
+                            ->removable()
+                            ->showOnExport()
+                            ->disk(config('filesystems.default'))
+                            ->dir('moonshine_users')
+                            ->allowedExtensions(['jpg', 'png', 'jpeg', 'gif']),
 
-                       Date::make(trans('moonshine::ui.resource.created_at'), 'created_at')
-                           ->format("d.m.Y")
-                           ->default(now()->toDateTimeString())
-                           ->sortable()
-                           ->hideOnForm()
-                           ->showOnExport(),
+                        Date::make(trans('moonshine::ui.resource.created_at'), 'created_at')
+                            ->format("d.m.Y")
+                            ->default(now()->toDateTimeString())
+                            ->sortable()
+                            ->hideOnForm()
+                            ->showOnExport(),
 
-                       Email::make(trans('moonshine::ui.resource.email'), 'email')
-                           ->sortable()
-                           ->showOnExport()
-                           ->required(),
-                   ]),
+                        Email::make(trans('moonshine::ui.resource.email'), 'email')
+                            ->sortable()
+                            ->showOnExport()
+                            ->required(),
+                    ]),
 
-                   Block::make(trans('moonshine::ui.resource.change_password'), [
-                       Password::make(trans('moonshine::ui.resource.password'), 'password')
-                           ->customAttributes(['autocomplete' => 'new-password'])
-                           ->hideOnIndex()
-                           ->hideOnExport()
-                           ->hideOnDetail()
-                           ->eye(),
+                    Block::make(trans('moonshine::ui.resource.change_password'), [
+                        Password::make(trans('moonshine::ui.resource.password'), 'password')
+                            ->customAttributes(['autocomplete' => 'new-password'])
+                            ->hideOnIndex()
+                            ->hideOnExport()
+                            ->hideOnDetail()
+                            ->eye(),
 
-                       PasswordRepeat::make(trans('moonshine::ui.resource.repeat_password'), 'password_repeat')
-                           ->customAttributes(['autocomplete' => 'confirm-password'])
-                           ->hideOnIndex()
-                           ->hideOnExport()
-                           ->hideOnDetail()
-                           ->eye(),
-                   ]),
-               ]),
+                        PasswordRepeat::make(trans('moonshine::ui.resource.repeat_password'), 'password_repeat')
+                            ->customAttributes(['autocomplete' => 'confirm-password'])
+                            ->hideOnIndex()
+                            ->hideOnExport()
+                            ->hideOnDetail()
+                            ->eye(),
+                    ]),
+                ]),
             ]),
         ];
     }
@@ -106,10 +107,10 @@ class MoonShineUserResource extends Resource
     {
         return [
             PermissionFormComponent::make('Permissions')
-                ->canSee(fn ($user) => $user->moonshine_user_role_id === MoonshineUserRole::DEFAULT_ROLE_ID),
+                ->canSee(fn($user) => $user->moonshine_user_role_id === MoonshineUserRole::DEFAULT_ROLE_ID),
 
             ChangeLogFormComponent::make('Change log')
-                ->canSee(fn ($user) => $user->moonshine_user_role_id === MoonshineUserRole::DEFAULT_ROLE_ID),
+                ->canSee(fn($user) => $user->moonshine_user_role_id === MoonshineUserRole::DEFAULT_ROLE_ID),
         ];
     }
 
@@ -125,7 +126,7 @@ class MoonShineUserResource extends Resource
                 'email',
                 Rule::unique('moonshine_users')->ignoreModel($item),
             ],
-            'password' => ! $item->exists
+            'password' => !$item->exists
                 ? 'required|min:6|required_with:password_repeat|same:password_repeat'
                 : 'sometimes|nullable|min:6|required_with:password_repeat|same:password_repeat',
         ];
@@ -149,17 +150,5 @@ class MoonShineUserResource extends Resource
         return [
             ExportAction::make(trans('moonshine::ui.export')),
         ];
-    }
-
-    public function resolveRoutes(): void
-    {
-        parent::resolveRoutes();
-
-        Route::prefix('resource')->group(function () {
-            Route::post(
-                "{$this->uriKey()}/{".$this->routeParam()."}/permissions",
-                PermissionController::class
-            )->name("{$this->routeNameAlias()}.permissions");
-        });
     }
 }
